// src/app/interfaces/orders-details.ts
export class OrdersDetails{
  orderContent: string;
  orderDeliveredDate: string;
  orderLocation: string;
  orderNumber: string;
  orderStartDate: string;
  orderState: string;
  orderStatus: string;
  orderChecked: number;
  orderOnEmployee: number;
  nameEmployee: string;
  lastNameEmployee: string;
  employeeContact:string;
}